forced GC in cleanup phase
